requirements.txt
pyspark==3.3.2
pandas==1.5.3
pynessie==0.51.1
findspark==2.0.1
delta-spark==2.2.0

backoff==2.2.1
black==22.12.0
boto==2.49.0
boto3==1.26.64
botocore==1.29.64
ratelimit==2.2.1
requests==2.28.2
s3transfer==0.6.0